#45 refactor: LoginTypeUseCase -> UserAuthInfoUseCase 타입명 , 하위 속성 타입 수정
- 키체인에서 인증 정보 자체를 가져옴에 따라 이름 및, 하위 속성 타입을 수정함
